Added elevation changes to ground floor
The floor height changes from a minimum of 1 to a maximum of 4. This elevation has a cooldown of 10 steps and has a 50% chance of occurring on the 11+ steps.

Issue: Hitting the side of the cliff should result in a game over, but implementing this might require changing the cell prefab.