Replaced MP3s with OGGs
Changed the music and made the file format OGG. This makes it so that the music loops more seamlessly since it is higher quality. Also mastered the music to match the sound effects volume. ALSO, made it so that loading the save file actually loads the volume for the music and sound effects.